validate load data inserts
