+++ sql/analyzer/validation_rules.go
@@ -834,75 +834,6 @@ func validateAggregations(ctx *sql.Context, a *Analyzer, n sql.Node, scope *plan
 	return n, transform.SameTree, validationErr
 }
 
-// validateInserts validates InsertInto nodes.
-func validateInserts(_ *sql.Context, _ *Analyzer, n sql.Node, _ *plan.Scope, _ RuleSelector) (sql.Node, transform.TreeIdentity, error) {
-	if _, ok := n.(*plan.TriggerExecutor); ok {
-		return n, transform.SameTree, nil
-	} else if _, ok := n.(*plan.CreateProcedure); ok {
-		return n, transform.SameTree, nil
-	}
-
-	return transform.Node(n, func(n sql.Node) (sql.Node, transform.TreeIdentity, error) {
-		// Skip any nodes that aren't an InsertInto
-		ii, ok := n.(*plan.InsertInto)
-		if !ok {
-			return n, transform.SameTree, nil
-		}
-
-		table := getResolvedTable(ii.Destination)
-
-		insertable, err := plan.GetInsertable(table)
-		if err != nil {
-			return n, transform.SameTree, err
-		}
-
-		if ii.IsReplace {
-			var ok bool
-			_, ok = insertable.(sql.ReplaceableTable)
-			if !ok {
-				return n, transform.SameTree, plan.ErrReplaceIntoNotSupported.New()
-			}
-		}
-
-		if len(ii.OnDupExprs) > 0 {
-			var ok bool
-			_, ok = insertable.(sql.UpdatableTable)
-			if !ok {
-				return n, transform.SameTree, plan.ErrOnDuplicateKeyUpdateNotSupported.New()
-			}
-		}
-
-		// normalize the column name
-		dstSchema := insertable.Schema()
-		columnNames := make([]string, len(ii.ColumnNames))
-		for i, name := range ii.ColumnNames {
-			columnNames[i] = strings.ToLower(name)
-		}
-
-		// If no columns are given and value tuples are not all empty, use the full schema
-		if len(columnNames) == 0 && existsNonZeroValueCount(ii.Source) {
-			columnNames = make([]string, len(dstSchema))
-			for i, f := range dstSchema {
-				columnNames[i] = f.Name
-			}
-		}
-
-		if len(ii.ColumnNames) > 0 {
-			err := validateColumns(table.Name(), columnNames, dstSchema, ii.Source)
-			if err != nil {
-				return n, transform.SameTree, err
-			}
-		}
-
-		err = validateValueCount(columnNames, ii.Source)
-		if err != nil {
-			return nil, transform.SameTree, err
-		}
-
-		return n, transform.SameTree, nil
-	})
-}
-
 // checkForAggregationFunctions returns an ErrAggregationUnsupported error if any aggregation
 // functions are found in the specified expressions.
 func checkForAggregationFunctions(exprs []sql.Expression) error {
